Rgeo fixes

+++ db/seeds.rb
@@ -1,4 +1,4 @@
-require 'pp'
+require 'awesome_print'
 require 'geokit'
 require 'rgeo'
 include Geokit::Geocoders
@@ -6,7 +6,7 @@ include Geokit::Geocoders
 # -------------------------------
 # Delete all unless in production
 # -------------------------------
-if Rails.env != "production"
+if Rails.env != "production" && ENV["force"]
   Location.delete_all
   Project.delete_all
   Event.delete_all
@@ -28,20 +28,19 @@ ring = factory.linear_ring([p1, p2, p3, p4, p1])
 world_geom = factory.polygon(ring)
 
 # World admin_level is set to -1 by convention
-# National: 1
-# Provincial: 2
+# National: 0
+# Provincial: 1
 world = Location.where(:name => 'World', :admin_level => -1).first_or_create(:geom => world_geom)
 
 # -------------------------------
 # Create users
 # -------------------------------
-gernot  = User.create(
-  :email                 => 'gernot.ritthaler@caritas.de',
+gernot  = User.where(:email => 'gernot.ritthaler@caritas.de').first_or_create(
   :password              => "caritas321",
   :password_confirmation => "caritas321",
   :role                  => 'admin'
 )
-gernot.settings = Settings.create!(:location => world)
+Settings.where(user_id: gernot.id).first_or_create!(:location => world)
 
 silvius = User.create(
   :email                 => 'prakkako@caritas.de',
@@ -49,7 +48,8 @@ silvius = User.create(
   :password_confirmation => "caritas321",
   :role                  => 'admin'
 )
-silvius.settings = Settings.create!(:location => world)
+#silvius.settings = Settings.create!(:location => world)
+Settings.where(user_id: silvius.id).first_or_create!(:location => world)
 
 publisher = User.create(
   :email                 => 'publisher@caritas.de',
@@ -57,7 +57,8 @@ publisher = User.create(
   :password_confirmation => "caritas123",
   :role                  => 'publisher'
 )
-publisher.settings = Settings.create!(:location => world)
+#publisher.settings = Settings.create!(:location => world)
+Settings.where(user_id: publisher.id).first_or_create!(:location => world)
 
 shoaib = User.create(
   :email                 => 'shoaib@nomad-labs.com',
@@ -65,7 +66,8 @@ shoaib = User.create(
   :password_confirmation => "caritas321",
   :role                  => 'admin'
 )
-shoaib.settings = Settings.create!(:location => world)
+#shoaib.settings = Settings.create!(:location => world)
+Settings.where(user_id: shoaib.id).first_or_create!(:location => world)
 
 public_relations = User.create(
   :email                 => 'public_relations@caritas.de',
@@ -73,7 +75,8 @@ public_relations = User.create(
   :password_confirmation => "caritas321",
   :role                  => 'public_relations'
 )
-public_relations.settings = Settings.create!(:location => world)
+#public_relations.settings = Settings.create!(:location => world)
+Settings.where(user_id: public_relations.id).first_or_create!(:location => world)
 
 oea = User.create(
   :email                 => 'oea@caritas.de',
@@ -81,7 +84,8 @@ oea = User.create(
   :password_confirmation => "caritas321",
   :role                  => 'public_relations'
 )
-oea.settings = Settings.create!(:location => world)
+#oea.settings = Settings.create!(:location => world)
+Settings.where(user_id: oea.id).first_or_create!(:location => world)
 
 monika = User.create(
   :email                 => 'monika.hoffmann@caritas.de',
@@ -89,55 +93,79 @@ monika = User.create(
   :password_confirmation => "caritas321",
   :role                  => 'public_relations'
 )
-monika.settings = Settings.create!(:location => world)
+#monika.settings = Settings.create!(:location => world)
+Settings.where(user_id: monika.id).first_or_create!(:location => world)
 
 # -----------------------------------
 # Load Locations data from shapefiles
 # -----------------------------------
 
 # load pakistan data
-puts %x[rake db:load_shapefile \
-  shapefile_path="#{Rails.root}/db/raw_data/pakistan/national_boundries_of_pakistan/pakistan_country_bdry.shp" \
-  name_field=NAME_1 \
-  description_field=descriptio \
-  admin_level=0 \
-  admin_email=#{gernot.email}]
-pakistan = Location.where(:name => "Pakistan").first
-puts "Created Pakistan id: #{pakistan.id}"
+if pakistan = Location.where(:name => "Pakistan").first
+  ap "Found Pakistan id: #{pakistan.id}"
+else
+  ap %x[rake db:load_shapefile \
+    shapefile_path="#{Rails.root}/db/raw_data/pakistan/national_boundries_of_pakistan/pakistan_country_bdry.shp" \
+    name_field=NAME_1 \
+    description_field=descriptio \
+    admin_level=0 \
+    parent_id=#{world.id} \
+    admin_email=#{gernot.email}]
+  pakistan = Location.where(:name => "Pakistan").first
+  ap "Created Pakistan id: #{pakistan.id}"
+end
 
-puts %x[rake db:load_shapefile \
-  shapefile_path="#{Rails.root}/db/raw_data/pakistan/provincial_boundaries_of_pakistan/provincial_boundaries_of_pakistan.shp" \
-  name_field=NAME_1 \
-  description_field=ENGTYPE_1 \
-  admin_level=1 \
-  parent_id=#{pakistan.id} \
-  admin_email=#{gernot.email}]
+if names_provinces_of_pakistan = pakistan.children.pluck(:name)
+  ap "Found Provinces of Pakistan"
+  ap names_provinces_of_pakistan
+  provinces_of_pakistan = pakistan.children.all
+else
+  ap %x[rake db:load_shapefile \
+    shapefile_path="#{Rails.root}/db/raw_data/pakistan/provincial_boundaries_of_pakistan/provincial_boundaries_of_pakistan.shp" \
+    name_field=NAME_1 \
+    description_field=ENGTYPE_1 \
+    admin_level=1 \
+    parent_id=#{pakistan.id} \
+    admin_email=#{gernot.email}]
+end
 
 # load Haiti data
-puts %x[rake db:load_shapefile \
-  shapefile_path="#{Rails.root}/db/raw_data/haiti/haiti_national/haiti_national_bdry.shp" \
-  name_field=COUNTRY \
-  description_field=COMMENT \
-  admin_level=0 \
-  admin_email=#{gernot.email}]
-haiti = Location.where(:name => "Republic of Haiti").first
-puts "Created Haiti id: #{haiti.id}"
+if haiti = Location.where(:name => "Republic of Haiti").first
+  ap "Found Haiti id: #{haiti.id}"
+else
+  ap %x[rake db:load_shapefile \
+    shapefile_path="#{Rails.root}/db/raw_data/haiti/haiti_national/haiti_national_bdry.shp" \
+    name_field=COUNTRY \
+    description_field=COMMENT \
+    admin_level=0 \
+    parent_id=#{world.id} \
+    admin_email=#{gernot.email}]
+  haiti = Location.where(:name => "Republic of Haiti").first
+  ap "Created Haiti id: #{haiti.id}"
+end
 
-puts %x[rake db:load_shapefile \
-  shapefile_path="#{Rails.root}/db/raw_data/haiti/Haiti_adm1_2000-2010.shp" \
-  name_field=ADM1_NAME \
-  description_field=COMMENT \
-  admin_level=1 \
-  parent_id=#{haiti.id} \
-  admin_email=#{gernot.email}]
+if names_provinces_of_haiti = haiti.children.pluck(:name)
+  ap "Found Provinces of Haiti"
+  ap names_provinces_of_haiti
+  provinces_of_haiti = haiti.children.all
+else
+  ap %x[rake db:load_shapefile \
+    shapefile_path="#{Rails.root}/db/raw_data/haiti/Haiti_adm1_2000-2010.shp" \
+    name_field=ADM1_NAME \
+    description_field=COMMENT \
+    admin_level=1 \
+    parent_id=#{haiti.id} \
+    admin_email=#{gernot.email}]
+end
 
 # -------------------------------
 # Create partners
 # -------------------------------
 partners = ['UNHCR', 'UNODC', 'ION', 'ILO', 'ICMC', 'JRS', 'COATNET', 'USG'].collect do |name|
-  p = Partner.new(:name => name)
+  p = Partner.where(:name => name).first_or_create!
   p.user = gernot
   p.save!
+  ap p.name
   p
 end
 
@@ -165,9 +193,10 @@ sectors = [
   'Livelihoods',
   'Partner Support'
 ].collect{ |name|
-  s = Sector.new(:name => name)
+  s = Sector.where(:name => name).first_or_create!
   s.user = gernot
   s.save!
+  ap s.name
   s
 }
 
@@ -217,16 +246,16 @@ proj_desc = "With help from partner agencies, Caritas Pakistan has since focused
       e = project.events.build(events_attributes[i])
       point = location.geom.envelope.centroid
 
-      width  = (country.geom.envelope.upper_corner.x - country.geom.envelope.lower_corner.x).abs/2
-      height = (country.geom.envelope.upper_corner.y - country.geom.envelope.lower_corner.y).abs/2
+      width  = (country.bbox_projected.max_point.x - country.bbox_projected.min_point.x).abs/2
+      height = (country.bbox_projected.max_point.y - country.bbox_projected.min_point.y).abs/2
 
-      point.with_z = true
-      point.x += width  * rand * 0.1 * [1,-1][rand(2)]
-      point.y += height * rand * 0.1 * [1,-1][rand(2)]
-      point.z = 0
-      res=GoogleGeocoder.reverse_geocode([point.y,point.x])
+      x = point.x + width  * rand * 0.1 * [1,-1][rand(2)]
+      y = point.y + height * rand * 0.1 * [1,-1][rand(2)]
+      z = point.z || 0
+      res=GoogleGeocoder.reverse_geocode([y,x])
       e.address = res.full_address
-      e.geom = point
+      event_location = Event::GEOM_FACTORY.point(x,y,z)
+      e.geom = event_location
       e.user = silvius
       e.save!
     end
